+++ crates/miden-lib/src/account/interface/test.rs
@@ -13,9 +13,9 @@ use miden_objects::note::{
     Note,
     NoteAssets,
     NoteExecutionHint,
-    NoteInputs,
     NoteMetadata,
     NoteRecipient,
+    NoteStorage,
     NoteTag,
     NoteType,
 };
@@ -293,7 +293,7 @@ fn test_basic_wallet_custom_notes() {
         end
     ";
     let note_script = ScriptBuilder::default().compile_note_script(compatible_source_code).unwrap();
-    let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
+    let recipient = NoteRecipient::new(serial_num, note_script, NoteStorage::default());
     let compatible_custom_note = Note::new(vault.clone(), metadata, recipient);
     assert_eq!(
         NoteAccountCompatibility::Maybe,
@@ -322,7 +322,7 @@ fn test_basic_wallet_custom_notes() {
     ";
     let note_script =
         ScriptBuilder::default().compile_note_script(incompatible_source_code).unwrap();
-    let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
+    let recipient = NoteRecipient::new(serial_num, note_script, NoteStorage::default());
     let incompatible_custom_note = Note::new(vault, metadata, recipient);
     assert_eq!(
         NoteAccountCompatibility::No,
@@ -382,7 +382,7 @@ fn test_basic_fungible_faucet_custom_notes() {
         end
     ";
     let note_script = ScriptBuilder::default().compile_note_script(compatible_source_code).unwrap();
-    let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
+    let recipient = NoteRecipient::new(serial_num, note_script, NoteStorage::default());
     let compatible_custom_note = Note::new(vault.clone(), metadata, recipient);
     assert_eq!(
         NoteAccountCompatibility::Maybe,
@@ -413,7 +413,7 @@ fn test_basic_fungible_faucet_custom_notes() {
     ";
     let note_script =
         ScriptBuilder::default().compile_note_script(incompatible_source_code).unwrap();
-    let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
+    let recipient = NoteRecipient::new(serial_num, note_script, NoteStorage::default());
     let incompatible_custom_note = Note::new(vault, metadata, recipient);
     assert_eq!(
         NoteAccountCompatibility::No,
@@ -499,7 +499,7 @@ fn test_custom_account_custom_notes() {
         .unwrap()
         .compile_note_script(compatible_source_code)
         .unwrap();
-    let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
+    let recipient = NoteRecipient::new(serial_num, note_script, NoteStorage::default());
     let compatible_custom_note = Note::new(vault.clone(), metadata, recipient);
     assert_eq!(
         NoteAccountCompatibility::Maybe,
@@ -526,7 +526,7 @@ fn test_custom_account_custom_notes() {
         .unwrap()
         .compile_note_script(incompatible_source_code)
         .unwrap();
-    let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
+    let recipient = NoteRecipient::new(serial_num, note_script, NoteStorage::default());
     let incompatible_custom_note = Note::new(vault, metadata, recipient);
     assert_eq!(
         NoteAccountCompatibility::No,
@@ -620,7 +620,7 @@ fn test_custom_account_multiple_components_custom_notes() {
         .unwrap()
         .compile_note_script(compatible_source_code)
         .unwrap();
-    let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
+    let recipient = NoteRecipient::new(serial_num, note_script, NoteStorage::default());
     let compatible_custom_note = Note::new(vault.clone(), metadata, recipient);
     assert_eq!(
         NoteAccountCompatibility::Maybe,
@@ -659,7 +659,7 @@ fn test_custom_account_multiple_components_custom_notes() {
         .unwrap()
         .compile_note_script(incompatible_source_code)
         .unwrap();
-    let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
+    let recipient = NoteRecipient::new(serial_num, note_script, NoteStorage::default());
     let incompatible_custom_note = Note::new(vault.clone(), metadata, recipient);
     assert_eq!(
         NoteAccountCompatibility::No,