+++ crates/miden-objects/src/account/component/mod.rs
@@ -131,8 +131,8 @@ impl AccountComponent {
     /// # Errors
     ///
     /// - If any of the component's storage entries cannot be transformed into a valid storage slot.
-    ///   This could be because the metadata is invalid, or storage values were not provided (or
-    ///   they are not of a valid type)
+    ///   This could be because the metadata is invalid, or storage items were not provided (or they
+    ///   are not of a valid type)
     pub fn from_template(
         template: &AccountComponentTemplate,
         init_storage_data: &InitStorageData,