+++ crates/miden-objects/src/errors.rs
@@ -38,7 +38,7 @@ use crate::{
     MAX_ACCOUNTS_PER_BATCH,
     MAX_INPUT_NOTES_PER_BATCH,
     MAX_INPUT_NOTES_PER_TX,
-    MAX_INPUTS_PER_NOTE,
+    MAX_NOTE_STORAGE_LENGTH,
     MAX_OUTPUT_NOTES_PER_TX,
 };
 
@@ -575,8 +575,8 @@ pub enum NoteError {
     NoteScriptDeserializationError(#[source] DeserializationError),
     #[error("note contains {0} assets which exceeds the maximum of {max}", max = NoteAssets::MAX_NUM_ASSETS)]
     TooManyAssets(usize),
-    #[error("note contains {0} inputs which exceeds the maximum of {max}", max = MAX_INPUTS_PER_NOTE)]
-    TooManyInputs(usize),
+    #[error("note contains {0} storage items which exceeds the maximum of {max}", max = MAX_NOTE_STORAGE_LENGTH)]
+    TooManyStorageItems(usize),
     #[error("note tag requires a public note but the note is of type {0}")]
     PublicNoteRequired(NoteType),
     #[error("{error_msg}")]