+++ crates/miden-objects/src/note/file.rs
@@ -149,10 +149,10 @@ mod tests {
         NoteAssets,
         NoteFile,
         NoteInclusionProof,
-        NoteInputs,
         NoteMetadata,
         NoteRecipient,
         NoteScript,
+        NoteStorage,
         NoteTag,
         NoteType,
     };
@@ -168,8 +168,8 @@ mod tests {
 
         let serial_num = Word::from([0, 1, 2, 3u32]);
         let script = NoteScript::mock();
-        let note_inputs = NoteInputs::new(vec![target.prefix().into()]).unwrap();
-        let recipient = NoteRecipient::new(serial_num, script, note_inputs);
+        let note_storage = NoteStorage::new(vec![target.prefix().into()]).unwrap();
+        let recipient = NoteRecipient::new(serial_num, script, note_storage);
 
         let asset = Asset::Fungible(FungibleAsset::new(faucet, 100).unwrap());
         let metadata = NoteMetadata::new(