+++ crates/miden-testing/tests/scripts/send_note.rs
@@ -10,9 +10,9 @@ use miden_objects::note::{
     Note,
     NoteAssets,
     NoteExecutionHint,
-    NoteInputs,
     NoteMetadata,
     NoteRecipient,
+    NoteStorage,
     NoteTag,
     NoteType,
     PartialNote,
@@ -46,7 +46,7 @@ async fn test_send_note_script_basic_wallet() -> anyhow::Result<()> {
     let assets = NoteAssets::new(vec![sent_asset]).unwrap();
     let note_script = ScriptBuilder::default().compile_note_script("begin nop end").unwrap();
     let serial_num = RpoRandomCoin::new(Word::from([1, 2, 3, 4u32])).draw_word();
-    let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
+    let recipient = NoteRecipient::new(serial_num, note_script, NoteStorage::default());
 
     let note = Note::new(assets.clone(), metadata, recipient);
     let partial_note: PartialNote = note.clone().into();
@@ -110,7 +110,7 @@ async fn test_send_note_script_basic_fungible_faucet() -> anyhow::Result<()> {
     )])?;
     let note_script = ScriptBuilder::default().compile_note_script("begin nop end").unwrap();
     let serial_num = RpoRandomCoin::new(Word::from([1, 2, 3, 4u32])).draw_word();
-    let recipient = NoteRecipient::new(serial_num, note_script, NoteInputs::default());
+    let recipient = NoteRecipient::new(serial_num, note_script, NoteStorage::default());
 
     let note = Note::new(assets.clone(), metadata, recipient);
     let partial_note: PartialNote = note.clone().into();